Move the output_store to a file_store file under output_stores
