+++ lib/dropdown/output_stores.rb
@@ -1,0 +1,2 @@
+require_relative 'output_stores/file_store'
+require_relative 'output_stores/dropbox_store'
