+++ lib/dropdown/output_stores/file_store.rb
